requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
pandas==2.1.1
numpy==1.26.0
python-multipart==0.0.6
pyyaml==6.0.1
openai>=1.0.0
